place e2e test to separate file
